feat(events): users can  add events
- fix validation issues
- prevent users from creating overlapping events

[Finishes #162705687]
