+++ app/models/concerns/date_time_helpers.rb
@@ -1,9 +1,14 @@
 module DateTimeHelpers
-  def time_overlaps?(times, start_time)
-    start_time = parse_datetime(start_time.to_s, :start_time)
-    times = times.select do |time|
-      # binding.pry
-      start_time < time
+  def time_overlaps?(existing_times, new_time)
+    start_time = parse_datetime(new_time[:start_time].to_s, :start_time)
+    end_time = parse_datetime(new_time[:end_time].to_s, :end_time)
+    times = existing_times.select do |time_group|
+      # time_group[0] -> start times | time_group[1] -> end times
+      start_time < time_group[1] && end_time > time_group[0]
+      # how condition works -> check if incoming start time is less than an existing
+      # end time for that center and then checks if the incoming end time if greater
+      # than an existing start time. If condition is true, the time_group is tracked by
+      # the select block
     end
     if times.count > 0
       true
@@ -22,12 +27,11 @@ module DateTimeHelpers
     end
   end
 
-  # Postman requests are failing. Fix tomorrow
   def parse_datetime(datetime, symbol)
     begin
       datetime = DateTime.parse(datetime)
-    rescue ArgumentError
-      # errors.add(symbol, e.to_s)
+    rescue ArgumentError => e
+      errors.add(symbol, e.to_s)
     end
     datetime
   end
